refactor(android): move chromecast things to own package
