+++ external_projects/bccm_player/android/src/main/kotlin/media/bcc/bccm_player/ExoPlayerController.kt
@@ -23,7 +23,7 @@ import kotlinx.coroutines.cancel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
-import media.bcc.bccm_player.CastMediaItemConverter.Companion.PLAYER_DATA_IS_LIVE
+import media.bcc.bccm_player.chromecast.CastMediaItemConverter.Companion.PLAYER_DATA_IS_LIVE
 import java.util.UUID
