+++ external_projects/bccm_player/android/src/main/kotlin/media/bcc/bccm_player/PlayerController.kt
@@ -5,9 +5,9 @@ import android.os.Bundle
 import androidx.media3.common.MediaItem
 import androidx.media3.common.MediaMetadata
 import androidx.media3.common.Player
-import media.bcc.bccm_player.CastMediaItemConverter.Companion.BCCM_EXTRAS
-import media.bcc.bccm_player.CastMediaItemConverter.Companion.PLAYER_DATA_IS_LIVE
-import media.bcc.bccm_player.CastMediaItemConverter.Companion.PLAYER_DATA_MIME_TYPE
+import media.bcc.bccm_player.chromecast.CastMediaItemConverter.Companion.BCCM_EXTRAS
+import media.bcc.bccm_player.chromecast.CastMediaItemConverter.Companion.PLAYER_DATA_IS_LIVE
+import media.bcc.bccm_player.chromecast.CastMediaItemConverter.Companion.PLAYER_DATA_MIME_TYPE
 
 
 abstract class PlayerController : Player.Listener {
