+++ external_projects/bccm_player/android/src/main/kotlin/media/bcc/bccm_player/chromecast/CastOptionsProvider.kt
@@ -1,10 +1,11 @@
-package media.bcc.bccm_player
+package media.bcc.bccm_player.chromecast
 
 import android.content.Context
 import com.google.android.gms.cast.framework.CastOptions
 import com.google.android.gms.cast.framework.OptionsProvider
 import com.google.android.gms.cast.framework.SessionProvider
 import com.google.android.gms.cast.framework.media.CastMediaOptions
+import media.bcc.bccm_player.R
 
 class CastOptionsProvider : OptionsProvider {
     override fun getCastOptions(context: Context): CastOptions {
