+++ external_projects/bccm_player/android/src/main/kotlin/media/bcc/bccm_player/chromecast/FLCastButton.kt
@@ -1,4 +1,4 @@
-package media.bcc.bccm_player
+package media.bcc.bccm_player.chromecast
 
 import android.content.Context
 import android.view.ContextThemeWrapper
@@ -8,6 +8,7 @@ import com.google.android.gms.cast.framework.CastButtonFactory
 import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
+import media.bcc.bccm_player.R
 
 class FLCastButtonFactory : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
     override fun create(context: Context, viewId: Int, args: Any?): PlatformView {
