--- external_projects/bccm_player/android/src/main/res/anim/dev2.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set  xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shareInterpolator="false">
-
-    <!--<translate
-        android:fromXDelta="100%" android:toXDelta="100%"
-        android:fromYDelta="100%" android:toYDelta="100%"
-        android:duration="1500" />
-    <scale
-        android:fromXScale="90%" android:toXScale="100%"
-        android:fromYScale="90%" android:toYScale="100%"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        android:duration="1500" />-->
-</set>
